Separate Foobar2ExampleContentProvider as its own ContentProvider
