+++ app/src/main/AndroidManifest.xml
@@ -18,8 +18,14 @@
         </activity>
 
         <provider
-            android:authorities="com.tehmou.rxandroidstores.example.provider.ExampleContentProvider"
-            android:name="com.tehmou.rxandroidstores.example.provider.ExampleContentProvider"
+            android:authorities="com.tehmou.rxandroidstores.example.provider.FoobarExampleContentProvider"
+            android:name="com.tehmou.rxandroidstores.example.provider.FoobarExampleContentProvider"
+            android:exported="false">
+        </provider>
+
+        <provider
+            android:authorities="com.tehmou.rxandroidstores.example.provider.Foobar2ExampleContentProvider"
+            android:name="com.tehmou.rxandroidstores.example.provider.Foobar2ExampleContentProvider"
             android:exported="false">
         </provider>
     </application>
